Extract plugin API code to separate file
